requirements.txt
aiohttp==1.0.5
astroid==2.0.4
async-timeout==3.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
discord.py==0.16.12
Flask==1.0.2
Flask-Login==0.4.1
gunicorn==19.9.0
idna==2.7
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
multidict==4.4.2
nbformat==4.4.0
numpy==1.15.2
oauthlib==2.1.0
plotly==3.3.0
psutil==5.4.7
pycparser==2.19
pylint==2.1.1
PyNaCl==1.0.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
retrying==1.3.3
six==1.11.0
SQLAlchemy==1.2.12
tabulate==0.8.2
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
websockets==3.4
Werkzeug==0.14.1
wrapt==1.10.11
youtube-dl==2018.10.5
